:sparkles: #20 feature : 문제 상세 조회 API

+++ src/main/java/com/algorithm/Algonote/repository/ProblemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-public interface ProblemRepository  extends JpaRepository <ProblemEntity, Long> {
+public interface ProblemRepository  extends JpaRepository <ProblemEntity, Integer> {
     List<ProblemEntity> findByUserid(String userid);
 }
